Fix to avoid CS0104
Invoking Statiq.App.Bootstrapper.AddDefaultNamespaces() adds the using
directives for *all* namespaces defined in the assembly into the C#
codes generated from a *.cshtml file. This results a CS0104 error when
multiple symbols have the same name but are defined in the different
namespaces. To avoid this behavior, the bootstrapper should be created
without Statiq.App.DefaultFeatures.Namespaces.